+++ semana16/todoList/src/utils/api_helper.ts
@@ -1,4 +1,5 @@
 import dayjs from 'dayjs'
+import { v4 as uuidv4 } from 'uuid'
 
 export function inputValido(input: string | number | undefined): boolean {
   if (input === undefined || input === 'undefined' || !input) {
@@ -15,6 +16,10 @@ export function todosValidos(inputs: Array<string>): boolean {
   return true
 }
 
-export function dataValida(dayjsDate: dayjs.Dayjs) {
+export function dataValida(dayjsDate: dayjs.Dayjs): boolean {
   return dayjsDate.isValid()
+}
+
+export function gerarId(): string {
+  return uuidv4()
 }
